ref(debugMeta): Rename folder and imports from DebugMeta-v2 with DebugMeta
