+++ static/app/components/events/interfaces/debugMeta/debugImage/index.tsx
