+++ static/app/components/events/interfaces/debugMeta/debugImage/processingIcon.tsx
