+++ static/app/components/events/interfaces/debugMeta/debugImageDetails/candidate/actions.tsx
