+++ static/app/components/events/interfaces/debugMeta/debugImageDetails/candidate/index.tsx
