+++ static/app/components/events/interfaces/debugMeta/debugImageDetails/candidate/information/index.tsx
