+++ static/app/components/events/interfaces/debugMeta/debugImageDetails/candidate/information/processingIcon.tsx
