+++ static/app/components/events/interfaces/debugMeta/debugImageDetails/candidate/status/index.tsx
