+++ static/app/components/events/interfaces/debugMeta/debugImageDetails/candidate/status/statusTooltip.tsx
