+++ static/app/components/events/interfaces/debugMeta/debugImageDetails/candidate/utils.tsx
