+++ static/app/components/events/interfaces/debugMeta/debugImageDetails/candidates.tsx
