+++ static/app/components/events/interfaces/debugMeta/debugImageDetails/generalInfo.tsx
