+++ static/app/components/events/interfaces/debugMeta/debugImageDetails/index.tsx
