+++ static/app/components/events/interfaces/debugMeta/debugImageDetails/utils.tsx
