+++ tests/js/test-balancer/jest-balance.json
@@ -72,7 +72,7 @@
   "/tests/js/spec/views/projectDetail/index.spec.tsx": 4283,
   "/tests/js/spec/views/performance/landing/queryBatcher.spec.tsx": 4263,
   "/tests/js/spec/views/organizationGroupDetails/groupEventEntries.spec.tsx": 4154,
-  "/tests/js/spec/components/events/interfaces/debugMeta-v2/imageDetailsCandidates.spec.tsx": 4093,
+  "/tests/js/spec/components/events/interfaces/debugMeta/imageDetailsCandidates.spec.tsx": 4093,
   "/tests/js/spec/views/eventsV2/tags.spec.jsx": 4000,
   "/tests/js/spec/components/group/sidebar.spec.jsx": 3963,
   "/tests/js/spec/views/accountDetail.spec.jsx": 3951,
@@ -457,7 +457,7 @@
   "/tests/js/spec/components/modals/emailVerificationModal.spec.js": 1437,
   "/tests/js/spec/components/resolutionBox.spec.tsx": 1435,
   "/tests/js/spec/utils/sessions.spec.tsx": 1435,
-  "/tests/js/spec/components/events/interfaces/debugMeta-v2/utils.spec.tsx": 1425,
+  "/tests/js/spec/components/events/interfaces/debugMeta/utils.spec.tsx": 1425,
   "/tests/js/spec/utils/getProjectsByTeams.spec.jsx": 1424,
   "/tests/js/spec/views/apiTokenRow.spec.jsx": 1423,
   "/tests/js/spec/components/search/sources/formSource.spec.jsx": 1423,
